admin/requirements.txt
Django==4.1
djangorestframework==3.13.1
mysqlclient==2.1.1
django-mysql==4.7.1
django-cors-headers==3.13.0
pika==1.3.0
Pillow==9.2.0